requirements.txt
pandas>=1.5.0
yfinance>=0.2.0
requests>=2.28.0
streamlit>=1.25.0
plotly>=5.15.0
python-dotenv>=1.0.0
numpy>=1.24.0 